requirements.txt
anthropic
boto3
beautifulsoup4
feedparser
flask
flask-cors
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-generativeai
googlesearch-python
gunicorn
huggingface_hub>=0.20
ics
langchain
langchain-community
langchain-core
langchain-openai
markdownify
matplotlib
moviepy
num2words
numpy
openai
openpyxl
pandas
pillow
pydub
PyMuPDF
PyPDF2
python-docx
python-dotenv
python-pptx
python-dateutil
pdfplumber
pytesseract
requests
requests-oauthlib
scikit-learn
sendgrid
transformers
tweepy
elevenlabs==0.2.26
